[service] Move proposal calculation to manager
